apps/receipt/requirements.txt
Flask
python-dotenv
azure-storage-blob
azure-identity
azure-monitor-opentelemetry
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc